requirements.txt
Flask>=2.0
requests>=2.28
beautifulsoup4>=4.11
lxml>=4.9
SQLAlchemy>=1.4
WeasyPrint>=57.0
APScheduler>=3.9
pytest>=7.0
pdfminer.six>=20221105
